+++ src/main/kotlin/com/example/days/domain/report/model/UserReport.kt
@@ -2,7 +2,6 @@ package com.example.days.domain.report.model
 
 import com.example.days.domain.user.model.Status
 import com.example.days.domain.user.model.User
-import com.example.days.domain.user.model.UserRole
 import com.example.days.global.entity.BaseEntity
 import jakarta.persistence.*
 import org.hibernate.annotations.OnDelete
