+++ src/main/kotlin/com/example/days/global/common/exception/common/ExistException.kt
@@ -41,7 +41,4 @@ data class NotHaveSearchException(val title: String) : RuntimeException(
     "검색 결과가 존재하지 않습니다."
 )
 
-data class NotReportStatusException(val status: String) : RuntimeException(
-    "SPAM, HARMFUL, ILLEGAL 중에 선택해서 입력해주길 바랍니다!"
-)
